requirements.txt
# backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.1
python-dotenv==1.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
sentry-sdk==1.39.1
pydantic==2.6.0
boto3==1.34.0
pymongo==4.5.0